Display geometry fields in the appropriate coordinate system (4326)
- We created a custom django admin widget based on GeoDjango
  OpenLayersWidget and OSMWidget
- The responsability to convert between two coordinates system is now on
  JS side instead of being done in a mix of python and JavaScript
