+++ carting/templates/admin/custom-openlayers.html
@@ -1,0 +1,6 @@
+{% extends "gis/openlayers-osm.html" %}
+
+{% block options %}{{ block.super }}
+options['map_epsg'] = '{{ map_epsg }}';
+options['dataset_epsg'] = '{{ dataset_epsg }}';
+{% endblock %}
